requirements.txt
fastapi
uvicorn
transformers
torch
accelerate
pydantic-settings
bitsandbytes
accelerate
peft
python-dotenv